+++ src/command_modules/azure-cli-appservice/HISTORY.rst
@@ -3,6 +3,10 @@
 Release History
 ===============
 
+0.1.37
+++++++
+* Minor fixes
+
 0.1.36
 ++++++
 * webapp/functionapp: Adding support for disabling identity az webapp identity remove. Preview tag removed for Identity feature.